requirements.txt
Flask
flask-jwt-extended
flask-bcrypt
mysql-connector-python
geopy
